+++ spec/dummy/app/controllers/inertia_server_head_ssr_controller.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class InertiaServerHeadSsrController < ApplicationController
+  inertia_config(server_head: true, layout: 'meta', ssr_enabled: true, version: '1.0')
+
+  def basic
+    render inertia: 'TestComponent', meta: [
+      { tag_name: 'title', inner_content: 'The Inertia title' }
+    ]
+  end
+end